wip: swarm middleware
