+++ libs/deepagents/deepagents/middleware/subagents.py
@@ -1,10 +1,7 @@
 """Middleware for providing subagents to an agent via a `task` tool."""
 
-import asyncio
-import json
-import uuid
 from collections.abc import Awaitable, Callable, Sequence
-from typing import Annotated, Any, NotRequired, TypedDict, cast
+from typing import Any, NotRequired, TypedDict, cast
 
 from langchain.agents import create_agent
 from langchain.agents.middleware import HumanInTheLoopMiddleware, InterruptOnConfig
@@ -14,11 +11,9 @@ from langchain_core.language_models import BaseChatModel
 from langchain_core.messages import HumanMessage, ToolMessage
 from langchain_core.runnables import Runnable
 from langchain_core.tools import StructuredTool
-from langgraph.errors import GraphInterrupt
 from langgraph.types import Command
-from pydantic import BaseModel, Field, TypeAdapter, ValidationError
+from pydantic import BaseModel, Field
 
-from deepagents.backends.composite import CompositeBackend
 from deepagents.backends.protocol import BackendFactory, BackendProtocol
 from deepagents.middleware._utils import append_to_system_message
 from deepagents.middleware.permissions import FilesystemPermission
@@ -295,67 +290,6 @@ When NOT to use the task tool:
 - Remember to use the `task` tool to silo independent tasks within a multi-part objective.
 - You should use the `task` tool whenever you have a complex task that will take multiple steps, and is independent from other tasks that the agent needs to complete. These agents are highly competent and efficient."""  # noqa: E501
 
-_SWARM_SYSTEM_PROMPT_ADDITION = """
-
-## `swarm` (parallel subagent execution)
-
-Use `swarm` to fan out many independent tasks across multiple subagents and aggregate their results.
-
-### When to use swarm
-
-**Trigger condition**: Use swarm when the input contains too much data to process in a single pass. Indicators: the file or dataset contains hundreds of items that each need individual analysis, or exceeds ~500 lines. When in doubt, check the size and prefer swarm over attempting to process a large input inline.
-
-Also use `swarm` when:
-- A task requires applying intelligence to each item in a large collection
-- Work can be decomposed into many independent, parallel subtasks
-
-Use `task` instead when:
-- You have a small number of independent subtasks
-- Each subtask depends on the result of a previous one
-- The work is exploratory or adaptive
-
-### How to use swarm
-
-Before calling swarm, understand what you're working with. Explore the data to learn its structure, format, and content using whatever tools are available. The goal is to write task descriptions detailed enough that each subagent can execute without needing to figure anything out on its own.
-
-Once you understand the data:
-
-1. **Generate tasks.** Write a generation script via `execute` that produces a `tasks.jsonl` file — one JSON object per line, each with `id`, `description`, and optional `subagent_type`. Each task should be a self-contained unit of work. **Prefer many small tasks over few large ones** — all tasks run in parallel, so 50 small tasks finish in roughly the same wall-clock time as 5 large ones. When splitting a file, aim for **30-60 lines** per chunk.
-2. **Call swarm.** Pass the path to your `tasks.jsonl` file.
-3. **Aggregate results.** Write an aggregation script via `execute` that reads `<results_dir>/results.jsonl` and combines the subagent outputs into a final answer.
-
-### Task description quality
-
-Each subagent receives **only its task description** — no other context. The quality of your descriptions determines the quality of swarm results. Invest time upfront to get them right.
-
-Good task descriptions are **prescriptive**: they tell the subagent the data format, the processing logic, the exact range of data to work on, and the expected output format. The subagent should not need to explore or interpret — just execute.
-
-When subagent results need to be aggregated (counting, classification, extraction), instruct each subagent to respond with **structured JSON only** — no explanations, no tables, just the JSON object. Include the exact output schema in the task description.
-
-### Error handling
-
-Each task runs exactly once — there are no automatic retries. If some tasks fail, the swarm summary includes a `failed_tasks` array with each failed task's ID and error message. Use this to decide:
-- **Retry via swarm**: generate a new tasks.jsonl targeting just the failures (with modifications) and call swarm again.
-- **Retry individually**: use `task` for a small number of failures.
-- **Proceed with partial results**: aggregate what completed and skip the rest.
-
-### Important: one swarm call per question
-
-**Never re-run swarm to verify or cross-check results.** Swarm is expensive — treat the first run's per-task outputs as authoritative. If you need to validate, do it in the aggregation script (e.g., check that each chunk returned the expected number of items). Do not generate a second tasks.jsonl or call swarm again for the same question.
-
-### Decomposition patterns
-
-**Flat fan-out**: Split a dataset into equal chunks. All tasks are identical in structure.
-Good for: large files, classification, extraction.
-
-**One-per-item**: One task per discrete unit (file, document, URL).
-Good for: summarizing collections, processing independent documents.
-
-**Dimensional**: Multiple tasks examine the same input from different angles.
-Good for: code review, multi-criteria evaluation."""  # noqa: E501
-
-TASK_SYSTEM_PROMPT = _TASK_ONLY_SYSTEM_PROMPT + _SWARM_SYSTEM_PROMPT_ADDITION
-
 
 DEFAULT_GENERAL_PURPOSE_DESCRIPTION = "General-purpose agent for researching complex questions, searching for files and content, and executing multi-step tasks. When you are searching for a keyword or file and are not confident that you will find the right match in the first few tries use this agent to perform the search for you. This agent has access to all tools as the main agent."  # noqa: E501
 
@@ -469,121 +403,6 @@ def _build_task_tool(  # noqa: C901
     )
 
 
-DEFAULT_SWARM_CONCURRENCY = 10
-"""Default concurrency limit for swarm execution."""
-
-MAX_SWARM_CONCURRENCY = 50
-"""Maximum allowed concurrency."""
-
-SWARM_TASK_TIMEOUT_SECONDS = 300
-"""Per-task timeout in seconds."""
-
-SWARM_TOOL_DESCRIPTION = """Execute a batch of independent tasks in parallel across multiple subagents.
-
-## Workflow
-
-1. Write a generation script via `execute` that produces a tasks.jsonl file with one JSON object per line:
-   ```json
-   {{"id": "chunk_0", "description": "Read lines 1-100 of data.txt. Process each item. Return JSON results.", "subagent_type": "general-purpose"}}
-   {{"id": "chunk_1", "description": "Read lines 101-200 of data.txt. Process each item. Return JSON results.", "subagent_type": "general-purpose"}}
-   ```
-2. Call `swarm` with the path to the tasks.jsonl file.
-3. The tool returns a JSON summary with `total`, `completed`, `failed`, and `results_dir`.
-   Results are written to `<results_dir>/results.jsonl` — each line is the original task enriched with `status`, `result`, and/or `error` fields.
-4. Write an aggregation script via `execute` that reads `<results_dir>/results.jsonl` and combines the outputs.
-
-## tasks.jsonl fields
-
-- "id" (string, required): unique task identifier
-- "description" (string, required): complete, self-contained prompt — the subagent receives NOTHING else
-- "subagent_type" (string, optional): which subagent to use (default: "general-purpose")
-
-## After execution
-
-The tool returns:
-```json
-{{"total": 20, "completed": 19, "failed": 1,
-  "results_dir": "swarm_runs/<uuid>",
-  "failed_tasks": [{{"id": "chunk_5", "error": "timed out"}}]}}
-```
-
-Each task runs exactly once — there are no automatic retries. Use the `failed_tasks` array to decide how to handle failures.
-
-Available subagent types: {available_agents}
-"""
-
-
-class SwarmTaskSpec(TypedDict):
-    """A single task line in a ``tasks.jsonl`` file.
-
-    Fields:
-        id: Unique task identifier (must be unique within the task list).
-        description: Complete, self-contained prompt for the subagent.
-        subagent_type: Which subagent to dispatch to. Defaults to
-            ``"general-purpose"`` when omitted.
-    """
-
-    id: str
-    description: str
-    subagent_type: NotRequired[str]
-
-
-class ParsedSwarmConfig(TypedDict):
-    """Result of parsing a tasks JSONL file."""
-
-    tasks: list[SwarmTaskSpec] | None
-    error: str | None
-
-
-def _parse_tasks_jsonl(content: str) -> ParsedSwarmConfig:
-    """Parse and validate a ``tasks.jsonl`` string into task specs.
-
-    Validates that each line is valid JSON with required fields, that all
-    task IDs are unique, and that at least one task is present.
-
-    Args:
-        content: Raw JSONL string (one JSON object per line).
-
-    Returns:
-        Parsed config with task list or error message.
-    """
-    lines = [line for line in content.split("\n") if line.strip()]
-    if not lines:
-        return {"tasks": None, "error": "tasks.jsonl is empty. The generation script must write at least one task."}
-
-    tasks: list[SwarmTaskSpec] = []
-    seen_ids: set[str] = set()
-    errors: list[str] = []
-
-    adapter = TypeAdapter(SwarmTaskSpec)
-    for idx, line in enumerate(lines):
-        line_number = idx + 1
-        try:
-            parsed = json.loads(line)
-        except json.JSONDecodeError:
-            errors.append(f"Line {line_number}: invalid JSON")
-            continue
-
-        try:
-            task = adapter.validate_python(parsed)
-        except ValidationError as e:
-            messages = [issue["msg"] for issue in e.errors()]
-            errors.append(f"Line {line_number}: {', '.join(messages)}")
-            continue
-
-        if task["id"] in seen_ids:
-            errors.append(f'Line {line_number}: duplicate task id "{task["id"]}"')
-            continue
-
-        seen_ids.add(task["id"])
-        tasks.append(task)
-
-    if errors:
-        return {"tasks": None, "error": "tasks.jsonl validation failed:\n" + "\n".join(errors)}
-
-    return {"tasks": tasks, "error": None}
-
-
 class SubAgentMiddleware(AgentMiddleware[Any, ContextT, ResponseT]):
     """Middleware for providing subagents to an agent via a `task` tool.
 
@@ -637,9 +456,8 @@ class SubAgentMiddleware(AgentMiddleware[Any, ContextT, ResponseT]):
         *,
         backend: BackendProtocol | BackendFactory,
         subagents: Sequence[SubAgent | CompiledSubAgent],
-        system_prompt: str | None = TASK_SYSTEM_PROMPT,
+        system_prompt: str | None = _TASK_ONLY_SYSTEM_PROMPT,
         task_description: str | None = None,
-        enable_swarm: bool = False,
     ) -> None:
         """Initialize the `SubAgentMiddleware`."""
         super().__init__()
@@ -650,12 +468,9 @@ class SubAgentMiddleware(AgentMiddleware[Any, ContextT, ResponseT]):
         self._backend = backend
         self._subagents = subagents
         subagent_specs = self._get_subagents()
+        self.subagent_specs = subagent_specs
 
         task_tool = _build_task_tool(subagent_specs, task_description)
-
-        # Use task-only prompt when swarm is disabled
-        if not enable_swarm and system_prompt == TASK_SYSTEM_PROMPT:
-            system_prompt = _TASK_ONLY_SYSTEM_PROMPT
 
         # Build system prompt with available agents
         if system_prompt and subagent_specs:
@@ -664,165 +479,7 @@ class SubAgentMiddleware(AgentMiddleware[Any, ContextT, ResponseT]):
         else:
             self.system_prompt = system_prompt
 
-        tools = [task_tool]
-        if enable_swarm and backend is not None:
-            tools.append(self._build_swarm_tool(subagent_specs))
-        self.tools = tools
-
-    def _build_swarm_tool(self, subagents: list[_SubagentSpec]) -> BaseTool:  # noqa: C901, PLR0915
-        """Create the `swarm` tool.
-
-        The `swarm` tool reads a JSON config file containing a list of task specs and
-        runs those tasks in parallel across subagents.
-
-        Each task result is written to `<output_dir>/<task_id>.txt` via the configured
-        backend.
-
-        Args:
-            subagents: The available subagents that `swarm` may dispatch to.
-
-        Returns:
-            A structured tool named `swarm`.
-        """
-        backend = self._backend
-        if backend is None:
-            msg = "backend is required"
-            raise ValueError(msg)
-
-        subagent_graphs: dict[str, Runnable] = {spec["name"]: spec["runnable"] for spec in subagents}
-        subagent_description_str = "\n".join(f"- {s['name']}: {s['description']}" for s in subagents)
-        description = SWARM_TOOL_DESCRIPTION.format(available_agents=subagent_description_str)
-
-        async def _run_single_task(
-            task: SwarmTaskSpec,
-            subagent: Runnable,
-            parent_state: dict[str, Any],
-        ) -> dict[str, Any]:
-            """Run a single task with a timeout. Returns a lean result dict."""
-            subagent_state = dict(parent_state)
-            subagent_state["messages"] = [HumanMessage(content=task["description"])]
-            subagent_type = task.get("subagent_type", "general-purpose")
-
-            try:
-                result = await asyncio.wait_for(
-                    subagent.ainvoke(subagent_state),
-                    timeout=SWARM_TASK_TIMEOUT_SECONDS,
-                )
-                messages = result.get("messages", [])
-                text = messages[-1].text.rstrip() if messages and messages[-1].text else ""
-                return {"id": task["id"], "subagent_type": subagent_type, "status": "completed", "result": text}
-            except Exception as exc:  # noqa: BLE001
-                return {"id": task["id"], "subagent_type": subagent_type, "status": "failed", "error": str(exc)}
-
-        async def _execute_swarm(
-            tasks: list[SwarmTaskSpec],
-            parent_state: dict[str, Any],
-            effective_concurrency: int,
-        ) -> list[dict[str, Any]]:
-            """Dispatch all tasks in parallel under a concurrency semaphore.
-
-            Each task runs exactly once — there are no retries. The
-            orchestrator owns error recovery.
-            """
-            semaphore = asyncio.Semaphore(effective_concurrency)
-
-            for task in tasks:
-                subagent_type = task.get("subagent_type", "general-purpose")
-                if subagent_type not in subagent_graphs:
-                    allowed = ", ".join(f'"{k}"' for k in subagent_graphs)
-                    msg = f'Task "{task["id"]}" references unknown subagent_type "{subagent_type}". Available: {allowed}'
-                    raise ValueError(msg)
-
-            async def run_with_semaphore(task: SwarmTaskSpec) -> dict[str, Any]:
-                subagent = subagent_graphs[task.get("subagent_type", "general-purpose")]
-                async with semaphore:
-                    return await _run_single_task(task, subagent, parent_state)
-
-            results = await asyncio.gather(
-                *[run_with_semaphore(t) for t in tasks],
-                return_exceptions=True,
-            )
-
-            final: list[dict[str, Any]] = []
-            for idx, raw_result in enumerate(results):
-                if isinstance(raw_result, GraphInterrupt):
-                    raise raw_result
-                if isinstance(raw_result, BaseException):
-                    subagent_type = tasks[idx].get("subagent_type", "general-purpose")
-                    final.append({"id": tasks[idx]["id"], "subagent_type": subagent_type, "status": "failed", "error": str(raw_result)})
-                else:
-                    final.append(raw_result)
-            return final
-
-        async def aswarm(
-            tasks_path: Annotated[str, "Path to the tasks.jsonl file produced by the generation script."],
-            runtime: ToolRuntime,
-            concurrency: Annotated[
-                int | None,
-                f"Maximum number of subagents running simultaneously. Default: {DEFAULT_SWARM_CONCURRENCY}, max: {MAX_SWARM_CONCURRENCY}.",
-            ] = None,
-        ) -> ToolMessage:
-            """Run swarm tasks."""
-            resolved_backend = backend(runtime) if callable(backend) else backend  # ty: ignore[call-top-callable]
-
-            responses = await resolved_backend.adownload_files([tasks_path])
-            response = responses[0]
-            if response.error:
-                return ToolMessage(
-                    content=f'Failed to read tasks file at "{tasks_path}". '
-                    f"Ensure the generation script writes the file to this exact path and try again.",
-                    status="error",
-                    tool_call_id=runtime.tool_call_id,
-                )
-            file_content = response.content
-            if not isinstance(file_content, bytes):
-                return ToolMessage(
-                    content=f"Content was expected to be bytes. Got {type(file_content)}.",
-                    status="error",
-                    tool_call_id=runtime.tool_call_id,
-                )
-            parsed = _parse_tasks_jsonl(file_content.decode("utf-8"))
-            if parsed["error"] is not None:
-                return ToolMessage(content=parsed["error"], status="error", tool_call_id=runtime.tool_call_id)
-            tasks = parsed["tasks"]
-            if tasks is None:
-                msg = "parsed swarm tasks unexpectedly missing"
-                raise AssertionError(msg)
-
-            parent_state = {k: v for k, v in runtime.state.items() if k not in _EXCLUDED_STATE_KEYS}
-
-            effective_concurrency = max(1, min(concurrency or DEFAULT_SWARM_CONCURRENCY, MAX_SWARM_CONCURRENCY))
-
-            task_results = await _execute_swarm(tasks, parent_state, effective_concurrency)
-
-            artifacts_root = resolved_backend.artifacts_root if isinstance(resolved_backend, CompositeBackend) else "/"
-            _root = artifacts_root.rstrip("/")
-            results_dir = f"{_root}/swarm_runs/{uuid.uuid4()}"
-            results_path = f"{results_dir}/results.jsonl"
-            results_content = "\n".join(json.dumps(r) for r in task_results) + "\n"
-            await resolved_backend.aupload_files([(results_path, results_content.encode("utf-8"))])
-
-            completed_count = sum(1 for r in task_results if r["status"] == "completed")
-            failed_results = [r for r in task_results if r["status"] == "failed"]
-            summary: dict[str, Any] = {
-                "total": len(task_results),
-                "completed": completed_count,
-                "failed": len(failed_results),
-                "results_dir": results_dir,
-                "failed_tasks": [{"id": r["id"], "error": r.get("error", "unknown")} for r in failed_results],
-            }
-
-            return ToolMessage(
-                content=json.dumps(summary),
-                status="success",
-                tool_call_id=runtime.tool_call_id,
-            )
-
-        return StructuredTool.from_function(
-            name="swarm",
-            coroutine=aswarm,
-            description=description,
-        )
+        self.tools = [task_tool]
 
     def _get_subagents(self) -> list[_SubagentSpec]:
         """Create runnable agents from specs.
